More 1.9.5
- Changed money back to double due to bug
- Removed mining info of backpack due to bug back
- Fixed settlers not going to right pricing
- Fixed quarry running multiple inventories at once

+++ civcraft/src/com/avrgaming/civcraft/command/plot/PlotCommand.java
@@ -216,7 +216,7 @@ public class PlotCommand extends CommandBase {
 		}
 		
 		try {
-			int price = Integer.valueOf(args[1]);
+			double price = Double.valueOf(args[1]);
 			tc.setForSale(true);
 			tc.setPrice(price);
 			tc.save();
