+++ civcraft/src/com/avrgaming/civcraft/config/ConfigTownLevel.java
@@ -28,9 +28,9 @@ import com.avrgaming.civcraft.main.CivLog;
 public class ConfigTownLevel {
 	public int level;
 	public String title;
-	public Integer upkeep;
+	public Double upkeep;
 	public int plots;
-	public Integer plot_cost;
+	public Double plot_cost;
 	public int tile_improvements;
 	
 	
@@ -41,9 +41,9 @@ public class ConfigTownLevel {
 			ConfigTownLevel town_level = new ConfigTownLevel();
 			town_level.level = (Integer)level.get("level");
 			town_level.title = (String)level.get("title");
-			town_level.upkeep = (Integer)level.get("upkeep");
+			town_level.upkeep = (Double)level.get("upkeep");
 			town_level.plots = (Integer)level.get("plots");
-			town_level.plot_cost = (Integer)level.get("plot_cost");
+			town_level.plot_cost = (Double)level.get("plot_cost");
 			town_level.tile_improvements = (Integer)level.get("tile_improvements");
 			
 			levels.put(town_level.level, town_level);
